Separate out symbol tests into their own module.
